requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.1.18
coverage==7.6.1
qrcode==8.2
Pillow==10.4.0
reportlab==4.0.9
openpyxl==3.1.2
